requirements.txt
aiofiles==22.1.0
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.80.0
frozenlist==1.3.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
yarl==1.8.1
